Rename polymer/feature-descriptor.ts -> polymer-core-feature.ts
